// backend/src/graphql/generated/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type RequireFields<T, K extends keyof T> = Omit<T, K> & { [P in K]-?: NonNullable<T[P]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
};

export type AuthPayload = {
  __typename?: 'AuthPayload';
  token: Scalars['String']['output'];
  user: User;
};

export type Comment = {
  __typename?: 'Comment';
  author: User;
  content: Scalars['String']['output'];
  createdAt: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  ticketId: Scalars['ID']['output'];
};

export type Holiday = {
  __typename?: 'Holiday';
  createdAt: Scalars['String']['output'];
  date: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
};

export type Mutation = {
  __typename?: 'Mutation';
  addComment: Comment;
  assignTicket: Ticket;
  changeTicketStatus: Ticket;
  createTicket: Ticket;
  login: AuthPayload;
  register: AuthPayload;
  resolveTicket: Ticket;
};
// ...
export type PageInfo = {
  __typename?: 'PageInfo';
  endCursor?: Maybe<Scalars['String']['output']>;
  hasNextPage: Scalars['Boolean']['output'];
};

export enum Priority {
  High = 'HIGH',
  Low = 'LOW',
  Medium = 'MEDIUM',
  Urgent = 'URGENT'
}

export type Query = {
  __typename?: 'Query';
  dashboard: TicketDashboard;
  holidays: Array<Holiday>;
  me?: Maybe<User>;
  ticket?: Maybe<Ticket>;
  tickets: TicketConnection;
  users: Array<User>;
};


export type QueryTicketArgs = {
  id: Scalars['ID']['input'];
};


export type QueryTicketsArgs = {
  assigneeId?: InputMaybe<Scalars['ID']['input']>;
  cursor?: InputMaybe<Scalars['String']['input']>;
  priority?: InputMaybe<Priority>;
  slaState?: InputMaybe<SlaState>;
  status?: InputMaybe<TicketStatus>;
  take?: InputMaybe<Scalars['Int']['input']>;
};


export type QueryUsersArgs = {
  role?: InputMaybe<UserRole>;
};

export type SlaInfo = {
  __typename?: 'SLAInfo';
  firstResponseDueAt: Scalars['String']['output'];
  firstResponseRemainingMinutes: Scalars['Int']['output'];
  firstResponseState: SlaState;
  resolutionDueAt: Scalars['String']['output'];
  resolutionRemainingMinutes: Scalars['Int']['output'];
  resolutionState: SlaState;
};

export enum SlaState {
  AtRisk = 'AT_RISK',
  Breached = 'BREACHED',
  OnTrack = 'ON_TRACK'
}

export type Ticket = {
  __typename?: 'Ticket';
  assignee?: Maybe<User>;
  comments: Array<Comment>;
  createdAt: Scalars['String']['output'];
  description: Scalars['String']['output'];
  firstResponseAt?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  priority: Priority;
  reporter: User;
  resolvedAt?: Maybe<Scalars['String']['output']>;
  sla: SlaInfo;
  status: TicketStatus;
  title: Scalars['String']['output'];
};

export type TicketConnection = {
  __typename?: 'TicketConnection';
  nodes: Array<Ticket>;
  pageInfo: PageInfo;
};

export type TicketDashboard = {
  __typename?: 'TicketDashboard';
  atRiskTickets: Scalars['Int']['output'];
  breachedTickets: Scalars['Int']['output'];
  inProgressTickets: Scalars['Int']['output'];
  openTickets: Scalars['Int']['output'];
};

export enum TicketStatus {
  Closed = 'CLOSED',
  InProgress = 'IN_PROGRESS',
  Open = 'OPEN',
  Resolved = 'RESOLVED'
}

export type User = {
  __typename?: 'User';
  createdAt: Scalars['String']['output'];
  email: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
  role: UserRole;
};

export enum UserRole {
  Agent = 'AGENT',
  Reporter = 'REPORTER'
}

export type WithIndex<TObject> = TObject & Record<string, any>;
export type ResolversObject<TObject> = WithIndex<TObject>;

export type ResolverTypeWrapper<T> = Promise<T> | T;


export type ResolverWithResolve<TResult, TParent, TContext, TArgs> = {
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> = ResolverFn<TResult, TParent, TContext, TArgs> | ResolverWithResolve<TResult, TParent, TContext, TArgs>;

export type ResolverFn<TResult, TParent, TContext, TArgs> = (
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo
) => Promise<TResult> | TResult;
// ...
export type ResolversTypes = ResolversObject<{
  AuthPayload: ResolverTypeWrapper<AuthPayload>;
  Boolean: ResolverTypeWrapper<Scalars['Boolean']['output']>;
  Comment: ResolverTypeWrapper<Comment>;
  Holiday: ResolverTypeWrapper<Holiday>;
  ID: ResolverTypeWrapper<Scalars['ID']['output']>;
  Int: ResolverTypeWrapper<Scalars['Int']['output']>;
  Mutation: ResolverTypeWrapper<{}>;
  PageInfo: ResolverTypeWrapper<PageInfo>;
  Priority: Priority;
  Query: ResolverTypeWrapper<{}>;
  SLAInfo: ResolverTypeWrapper<SlaInfo>;
  SLAState: SlaState;
  String: ResolverTypeWrapper<Scalars['String']['output']>;
  Ticket: ResolverTypeWrapper<Ticket>;
  TicketConnection: ResolverTypeWrapper<TicketConnection>;
  TicketDashboard: ResolverTypeWrapper<TicketDashboard>;
  TicketStatus: TicketStatus;
  User: ResolverTypeWrapper<User>;
  UserRole: UserRole;
}>;

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = ResolversObject<{
  AuthPayload: AuthPayload;
  Boolean: Scalars['Boolean']['output'];
  Comment: Comment;
  Holiday: Holiday;
  ID: Scalars['ID']['output'];
  Int: Scalars['Int']['output'];
  Mutation: {};
  PageInfo: PageInfo;
  Query: {};
  SLAInfo: SlaInfo;
  String: Scalars['String']['output'];
  Ticket: Ticket;
  TicketConnection: TicketConnection;
  TicketDashboard: TicketDashboard;
  User: User;
}>;
// ...
export type QueryResolvers<ContextType = GraphQLContext, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = ResolversObject<{
  dashboard?: Resolver<ResolversTypes['TicketDashboard'], ParentType, ContextType>;
  holidays?: Resolver<Array<ResolversTypes['Holiday']>, ParentType, ContextType>;
  me?: Resolver<Maybe<ResolversTypes['User']>, ParentType, ContextType>;
  ticket?: Resolver<Maybe<ResolversTypes['Ticket']>, ParentType, ContextType, RequireFields<QueryTicketArgs, 'id'>>;
  tickets?: Resolver<ResolversTypes['TicketConnection'], ParentType, ContextType, Partial<QueryTicketsArgs>>;
  users?: Resolver<Array<ResolversTypes['User']>, ParentType, ContextType, Partial<QueryUsersArgs>>;
}>;